Create Earth_MarsTransmissionDelay.py
This file contains the source code for a program that calculates the one way light lag and communications delay from Earth to Mars.  It can also display this information in various useful graphs including one that show orbital positions of the inner planets and one that shows how time delay changes over time.